use the official repo
Prior to this change, no official repo existed. Now that the vyatta binary packages have moved under the wireguard project umbrella, we can point to the official release repo, not to an individual contributors one.

With the release change, the naming scheme for the packages changed slightly and drops the "wireguard-" prefix.

This change adjusts the repo url and the expression used to identify the latest release accordingly. It also bumps the version.

Fixes #11